Move Application Gateway subresource ID parsing and validation to separate files
